[SUITE3-2136] Performance Analysis Of Threads - Statistic View
New load statistic metrics are created for all threads of
a process in class ProcessStatisticHelper.

If one or more of these metrics report a finite mean load
for the first time a new item is created as child of the
'Load' item of the process for each thread.

Additionally a child item 'Main Thread' is created for
the process itself in this case.

The child items show the load of their corresponding thread
and the 'Load' parent shows the sum of the load of all
threads.

Additional changes:

- The QSignalBlocker used to suppress gui updates after every
  call to setData() was moved to the update sub-functions
  to be able to enable signals if a new child item is added
  so that the statistic view gets updated.

Change-Id: Ie657a197d25405ed65f82de06fe46d06be178a40