[AM2INC-87] Import APP4MC Event Chains as Event Sequences
This commit implements the import of amalthea EventChains as
sm3::EventSequences. This allows for the translation of a
subset of amalthea EventChains which have no parallel paths.

To this end stub work methods for amalthea classes from the
Event Model and the Constraints Model were added to the
Converter.

Conversion code was added for amathea Events which have a
corresponding TraceEvent in the INCHRON model.

The work method for the amalthea EventChain uses the
method Converter::getAmaltheaEventSequence() to retrieve
the sequential list of amalthea Events for an amalthea
EventChain. This method throws a std::invalid_argument
with a speaking error message in case of any problem.

Issue: AM2INC-87
Change-Id: I36630c00b6d8144f89e47de34fb10b4f1fa3183c