[AM2INC-130] Complete Creation of EventGraphs
The proper predecessor/successor relations are now set for
the EventGraphTraceEvents which constitute the EventGraph.

The name of the EventGraphTraceEvent is now set to the
same name as the amalthea events (it was not set before).

The response event of an amalthea EventChain is now added
to the chain of events if the last event added by traversing
the child elements not already contain this event as last
step.

Change-Id: I836a7f236d22dbd8ce23fe296e986d48ba91eaaf
^AM2INC-130 fixed