[AM2INC-132] Amalthea Import of Event Chains with Joins fails
A recursive function Converter::fillEventGraph() has been implemented
which traverses the hierarchy of sub-EventChains of an amalthea
EventChain and adds all steps of a sub-chain except for the stimulus
and response events. Instead it returns the EventGraphNodes of
stimuli-successors and the response-predecessor, so that the correct
predecessor-/successor-relations can be set on a higher level
of the callstack.

This function now can treat sequential as well as parallel
sub-EventChains. For inner EventChains also EventSets are
supported but not for the top-level stimulus/response steps.

The import of EventChainTimingRequirements now also translates
the 'Type' attribute (Age - FirstLast, Reaction - FirstFirst) of
the EventChainLatencyConstraint (fixes AM2INC-131).

Change-Id: I34e9c0166c3a15b390bc62e909f1f00950990ef7
^AM2INC-131 fixed
^AM2INC-132 fixed