[AM2INC-158] Crash Translating democar-extended.amxmi
The democar-extended amalthea model contains ProcessPrototypes
which are used to define raw data of a task according to the
APP4MC documentation.

We do not translate them into any INCHRON ModelObject but we
have to prevent the traversal of children of these prototypes
because they may contain RunnableCalls or LabelAccesses which
are then parsed in the wrong context.

Change-Id: I6cf6f8ef3d08e6caa537ea03899c9a09673be26e